# Go: NewTensor & Value performance improvement