Considerable improvements in NewTensor & Value performance for non-String Tensors. 

```
name                           old time/op    new time/op    delta
Tensor/New/[150528]int32-16      1.95ms ± 1%    0.11ms ±14%   -94.12%  (p=0.000 n=8+7)
Tensor/New/[]float32-16          12.2ms ± 1%     0.8ms ±66%   -93.68%  (p=0.000 n=7+8)
Tensor/New/[][]float32-16        14.0ms ± 1%     1.2ms ±14%   -91.36%  (p=0.000 n=8+8)
Tensor/New/[][][]float32-16      13.9ms ± 1%     1.3ms ± 7%   -90.81%  (p=0.000 n=8+8)
Tensor/Value/[150528]int32-16     549µs ± 5%      61µs ± 2%   -88.83%  (p=0.000 n=8+8)
Tensor/Value/[]float32-16        11.9ms ± 0%     0.5ms ± 1%   -95.89%  (p=0.000 n=7+8)
Tensor/Value/[][]float32-16      14.9ms ± 1%     0.5ms ± 1%   -96.35%  (p=0.000 n=8+8)
Tensor/Value/[][][]float32-16    14.9ms ± 1%     0.5ms ± 2%   -96.33%  (p=0.000 n=8+8)

name                           old alloc/op   new alloc/op   delta
Tensor/New/[150528]int32-16       606kB ± 0%       0kB ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/New/[]float32-16          4.01MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/New/[][]float32-16        4.48MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/New/[][][]float32-16      4.48MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/Value/[150528]int32-16    1.21MB ± 0%    0.61MB ± 0%   -50.00%  (p=0.000 n=8+8)
Tensor/Value/[]float32-16        8.01MB ± 0%    4.01MB ± 0%   -50.00%  (p=0.000 n=8+8)
Tensor/Value/[][]float32-16      9.21MB ± 0%    4.25MB ± 0%   -53.82%  (p=0.000 n=8+8)
Tensor/Value/[][][]float32-16    9.23MB ± 0%    4.25MB ± 0%   -53.93%  (p=0.000 n=8+8)

name                           old allocs/op  new allocs/op  delta
Tensor/New/[150528]int32-16        4.00 ± 0%      3.00 ± 0%   -25.00%  (p=0.000 n=8+8)
Tensor/New/[]float32-16            4.00 ± 0%      3.00 ± 0%   -25.00%  (p=0.000 n=8+8)
Tensor/New/[][]float32-16         20.0k ± 0%      0.0k ± 0%   -99.98%  (p=0.000 n=8+8)
Tensor/New/[][][]float32-16       20.0k ± 0%      0.0k ± 0%   -99.98%  (p=0.000 n=8+8)
Tensor/Value/[150528]int32-16      7.00 ± 0%      2.00 ± 0%   -71.43%  (p=0.000 n=8+8)
Tensor/Value/[]float32-16          7.00 ± 0%      2.00 ± 0%   -71.43%  (p=0.000 n=8+8)
Tensor/Value/[][]float32-16       40.0k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/Value/[][][]float32-16     40.2k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=8+8)
```
Discussed in https://github.com/tensorflow/tensorflow/issues/36288. We avoid using binary.Write as this is both quite wordy and causes a large number of allocations with tensors of order 2 or above. What we do instead is just copy the memory directly when we can. Note since we're looking for native byte order for the machine copying memory is OK as it is in native byte order and Go primitive types are fundamentally the same as C primitive types as these are the machine's primitive types.