@philpearl 2020-02-13
I've got a follow-up PR making similar improvements for string tensors. NewTensor and Value take less than a tenth of the time and cause 1,000,000 times fewer allocations. I've parked it here for now: https://github.com/philpearl/tensorflow/pull/1

I feel I should note that I took interest in this because we encountered a real problem serving models with Go. Manipulating the Tensors caused so many allocations that the Go GC could not keep up and the memory usage exploded. We were very lucky it didn't cause a major outage.