@jhseu 2020-02-21
This was rolled back due to MemorySanitizer complaining about unsafe pointer arithmetic in TestSessionRunNeg from session_test.go. I haven't looked closely at it yet.