@jhseu 2020-02-21
Seems like it hit one of the throws in this function:
https://github.com/golang/go/blob/master/src/runtime/checkptr.go#L22