@philpearl 2020-02-22
I'm on vacation for another week. I'll take a look when I get back.

On Fri, 21 Feb 2020, 14:01 Jonathan Hseu, <notifications@github.com> wrote:

> Seems like it hit one of the throws in this function:
> https://github.com/golang/go/blob/master/src/runtime/checkptr.go#L22
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/tensorflow/tensorflow/pull/36578?email_source=notifications&email_token=AADPPXG5OSDBTTQVIDFDYLTREBFKPA5CNFSM4KR3XF62YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEMUHVQA#issuecomment-589855424>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADPPXCNKWMADMGZ5JFBEFTREBFKPANCNFSM4KR3XF6Q>
> .
>