@philpearl 2020-02-29
It looks like the new checkptr stuff in Go 1.14 (which is enabled with -race or -msan) doesn't cope with pointer arithmetic if the pointer is obtained from a slice via reflection. I can just turn checkptr off in the failing routine with a go:nocheckptr pragma. But I'll see if I can find a neater way.