@philpearl 2020-02-29
Have fixed in https://github.com/tensorflow/tensorflow/pull/37190.

Just FYI I'm developing on the v2.1 tag as the master branch doesn't build straightforwardly. Would it be possible to commit a proper set of generated files so `go build` and `go test` just work out of the box?