@jhseu 2020-03-10
Yeah, I proposed that here:
https://github.com/tensorflow/tensorflow/pull/36523

But it ended up being a bit controversial. The Go team suggested that it often causes issues due to conflicts between the generated protobufs from bazel and the checked-in versions. I don't have a great solution for now besides asking users to run `go generate` manually.