@mdempsky 2020-05-01
Hi, I wrote the checkptr feature. I found this issue searching GitHub for instances of "go:nocheckptr".

checkptr isn't expected to have false positives, but the pointer safety rules are somewhat subtle, so at times how to write code to satisfy it might be non-obvious. If you need help with this, feel free to add me as a reviewer.

In commit fa5ff875a5bb7eec732db332e4143340942ff855, there were two significant issues:

1. It's not safe to allocate variables of type `reflect.SliceHeader`. That type is only allowed to be used as a pointer to an actual slice-typed variable. You'd probably benefit from adding a helper function like:

        func setSlice(h *reflect.SliceHeader, ptr unsafe.Pointer, len, cap int) {
            h.Data = uintptr(ptr)
            h.Len = len
            h.Cap = cap
        }

   and using that everywhere you need to unsafely mutate a slice.

2. The result value from reflect.Value.Pointer must be immediately converted to unsafe.Pointer, before converting back again to uintptr for any pointer arithmetic. (See https://github.com/google/go-cmp/issues/167#issuecomment-546093202 for explanation why.)

   E.g., in setSliceInSlice, it needed to be `unsafe.Pointer(uintptr(unsafe.Pointer(slice.Pointer())) + (uintptr(index) * sliceSize))`.