@philpearl 2020-05-01
Thanks @mdempsky! That point about casting uintptr returns is particularly subtle to say the least.